+++ Changelog.md
@@ -3,6 +3,7 @@
 - Refactor
 - Unit Tests (Karma + Chai)
 - `track.options.looping` &rarr; `track.options.loop`
+- `track.tweenPan()` now uses `linearRampToValueAtTime()`
 
 ## 0.2.0